Refactor loadtest.go to use new Jamf Pro client configuration
