+++ go.mod
@@ -8,7 +8,7 @@ require (
 	github.com/aws/aws-sdk-go-v2/credentials v1.17.11
 	github.com/aws/aws-sdk-go-v2/feature/s3/manager v1.16.15
 	github.com/aws/aws-sdk-go-v2/service/s3 v1.53.1
-	github.com/deploymenttheory/go-api-http-client v0.1.30
+	github.com/deploymenttheory/go-api-http-client v0.1.32
 	github.com/mitchellh/mapstructure v1.5.0
 )
